+++ src/lib.rs
@@ -1,44 +1,54 @@
-#![deny(missing_debug_implementations)]
-#![deny(missing_docs)]
-#![deny(clippy::all)]
-#![cfg_attr(not(feature = "std"), no_std)]
-#![cfg_attr(feature = "unstable", feature(unsize, coerce_unsized))]
-
-//! A continuous fixed-length byte array with a specified alignment.
+//! Common utilities to work with aligned values and allocation.
 //!
 //! # Example
+//!
 //! ```
-//! use aligned_bytes::AlignedBytes;
+//! use aligned_utils::stack::Align8;
+//! let mut arr = Align8([1, 2, 3]);
+//! let bytes: &[u8] = &*arr;
+//! ```
+//!
+//! ```
+//! # #[cfg(feature="alloc")]
+//! # {
+//! use aligned_utils::bytes::AlignedBytes; // with feature "alloc"
 //! let mut bytes = AlignedBytes::new_zeroed(1024, 8);
 //! let buf: &mut [u8] = &mut *bytes;
+//! # }
 //! ```
+//!
 
+#![deny(
+    anonymous_parameters,
+    bare_trait_objects,
+    elided_lifetimes_in_paths,
+    missing_copy_implementations,
+    missing_debug_implementations,
+    missing_docs,
+    single_use_lifetimes,
+    trivial_casts,
+    trivial_numeric_casts,
+    unreachable_pub,
+    unstable_features,
+    unused_extern_crates,
+    unused_import_braces,
+    unused_qualifications,
+    unused_results,
+    clippy::all,
+    clippy::pedantic,
+    clippy::nursery,
+    clippy::cargo
+)]
+#![allow(box_pointers, clippy::module_name_repetitions)]
+#![no_std]
+
+#[cfg(feature = "alloc")]
 extern crate alloc;
 
-mod boxed;
-mod bytes;
+#[cfg(feature = "std")]
+extern crate std;
+
 pub mod stack;
 
-pub use self::boxed::AlignedBox;
-pub use self::bytes::AlignedBytes;
-
-use alloc::alloc::{handle_alloc_error, Layout};
-
-unsafe fn aligned_alloc(alloc: unsafe fn(Layout) -> *mut u8, size: usize, align: usize) -> *mut u8 {
-    let layout = match Layout::from_size_align(size, align) {
-        Ok(layout) => layout,
-        Err(_) => panic!("Invalid layout: size = {}, align = {}", size, align),
-    };
-
-    let ptr = alloc(layout);
-    if ptr.is_null() {
-        handle_alloc_error(layout);
-    }
-    debug_assert!(
-        (ptr as usize) % align == 0,
-        "pointer = {:p} is not a multiple of alignment = {}",
-        ptr,
-        align
-    );
-    ptr
-}
+#[cfg(feature = "alloc")]
+pub mod bytes;