+++ CHANGELOG.md
@@ -7,11 +7,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-[Unreleased]: https://github.com/datenlord/aligned-bytes/compare/v0.1.5...HEAD
+[Unreleased]: https://github.com/datenlord/aligned-utils/compare/v0.1.5...HEAD
+
+The project has been renamed to `aligned-utils`.
+
+## Added
+
++ Add `AlignedBytes::new_from_slice`
+
+## Changed
+
+* Move `AlignedBytes` to `crate::bytes::AlignedBytes`.
+
+## Removed
+
+- Remove `AlignedBox`
+
+## Fixed
+
++ Yank `v0.1.4`...`v0.1.5` due to a null pointer UB.
 
 ## [0.1.5] - 2020-10-18
 
-[0.1.5]: https://github.com/datenlord/aligned-bytes/compare/v0.1.4...v0.1.5
+[0.1.5]: https://github.com/datenlord/aligned-utils/compare/v0.1.4...v0.1.5
 
 The project has been moved from `Nugine/aligned-bytes` to `datenlord/aligned-bytes`.
 
@@ -21,7 +39,7 @@ The project has been moved from `Nugine/aligned-bytes` to `datenlord/aligned-byt
 
 ## [0.1.4] - 2020-09-05
 
-[0.1.4]: https://github.com/datenlord/aligned-bytes/compare/v0.1.3...v0.1.4
+[0.1.4]: https://github.com/datenlord/aligned-utils/compare/v0.1.3...v0.1.4
 
 ### Fixed
 
@@ -30,7 +48,7 @@ The project has been moved from `Nugine/aligned-bytes` to `datenlord/aligned-byt
 
 ## [0.1.3] - 2020-09-03
 
-[0.1.3]: https://github.com/datenlord/aligned-bytes/compare/v0.1.2...v0.1.3
+[0.1.3]: https://github.com/datenlord/aligned-utils/compare/v0.1.2...v0.1.3
 
 ### Added
 
@@ -43,7 +61,7 @@ The project has been moved from `Nugine/aligned-bytes` to `datenlord/aligned-byt
 
 ## [0.1.2] - 2020-08-24
 
-[0.1.2]: https://github.com/datenlord/aligned-bytes/compare/v0.1.1...v0.1.2
+[0.1.2]: https://github.com/datenlord/aligned-utils/compare/v0.1.1...v0.1.2
 
 ### Added
 
@@ -51,7 +69,7 @@ The project has been moved from `Nugine/aligned-bytes` to `datenlord/aligned-byt
 
 ## [0.1.1] - 2020-07-16
 
-[0.1.1]: https://github.com/datenlord/aligned-bytes/compare/v0.1.0...v0.1.1
+[0.1.1]: https://github.com/datenlord/aligned-utils/compare/v0.1.0...v0.1.1
 
 ### Added
 + `AlignedBytes` is now `Send + Sync`.
@@ -59,7 +77,7 @@ The project has been moved from `Nugine/aligned-bytes` to `datenlord/aligned-byt
 
 ## [0.1.0] - 2020-07-16
 
-[0.1.0]: https://github.com/datenlord/aligned-bytes/tree/v0.1.0
+[0.1.0]: https://github.com/datenlord/aligned-utils/tree/v0.1.0
 
 ### Added
 + `AlignedBytes` implementation.